Revert "fix language setting bug"
