+++ package.json
@@ -15,8 +15,6 @@
     "flow": "flow"
   },
   "dependencies": {
-    "@material-ui/core": "^3.9.3",
-    "@material-ui/icons": "^3.0.2",
     "antd": "^3.16.1",
     "aws-amplify": "^1.1.24",
     "aws-amplify-react": "^2.3.4",
