+++ src/components/business_profile/postJob.js
@@ -6,14 +6,14 @@ import JobItem from './jobItem';
 const {Content} = Layout; 
 const postJob = (props) => {
     return (
-        
+        <container>
         <Layout>
             <Content>
                 <JobItem jobs={props.jobList} />
             </Content>
         </Layout>
 
-   
+    </container>
     )
 }
 export default postJob;
