+++ src/components/form/postJob.js
@@ -4,9 +4,8 @@ import React from "react"
 //import Layout from "../components/layout"
 import { Form, Icon, Input, Button, Tooltip, DatePicker, Select } from 'antd';
 import { I18n } from 'aws-amplify';
-import dict from "../dictionary/dictionary";
 //import "../style/postJob.css";
-
+import dict from "../dictionary/dictionary"
 const Option = Select.Option;
 const { TextArea } = Input;
 
@@ -14,12 +13,14 @@ class PostJob extends React.Component {
 
     state = {
         lan: window.localStorage.getItem('lan')
-    } 
+    }
+    
     handleSubmit = () => {
 
     }
 
     render() {
+        console.log("language",window.localStorage.getItem('lan'));
         I18n.putVocabularies(dict);
         I18n.setLanguage(this.state.lan);
         return (
