+++ src/components/navBar.js
@@ -4,6 +4,7 @@ import { getUser, isLoggedIn, logout } from '../services/auth';
 import { Auth } from "aws-amplify";
 import { Menu, Icon, Avatar, Button } from 'antd';
 import { I18n } from 'aws-amplify';
+import dict from "./dictionary/dictionary";
 
 const navBar = (props) => {
   const state = {
@@ -50,7 +51,7 @@ const navBar = (props) => {
       </Menu.Item>   
       <Menu.Item>
         <Icon type="profile" theme="twoTone" />User Profile
-        <Link to="/app/user-profile/userID"></Link>
+        <Link to="/app/user-profile"></Link>
       </Menu.Item>   
       {!isLoggedIn() ? (
       <Menu.Item key="register">
@@ -59,12 +60,14 @@ const navBar = (props) => {
       </Menu.Item>):null}
       <Menu.Item>
         <Button ghost="true" onClick={() => {
+          I18n.putVocabularies(dict);
           window.localStorage.setItem('lan', 'es');
           window.location.reload();
         }}>ENGLISH</Button>
       </Menu.Item>
       <Menu.Item>
         <Button ghost="true" onClick={() => {
+          I18n.putVocabularies(dict);
           window.localStorage.setItem('lan', 'ch');
           window.location.reload();
         }}>中文</Button>
